% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{get_aes_values}
\alias{get_aes_values}
\title{Get project-specific aesthetics values}
\usage{
get_aes_values(aesthetic, parameterization)
}
\description{
For instance, parameterization-dependent color and fills
}
