% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{get_col_types}
\alias{get_col_types}
\title{Get column types for extracting data from csv files}
\usage{
get_col_types(file_type)
}
\arguments{
\item{file_type}{Type of file to be read}
}
\description{
File type can be any of the following:
}
\details{
trial_data_ds - tidied trial-level data of defer-speedup task
trial_data_dd - tidied trial-level data of date-delay task
}
