% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_ips_calibration}
\alias{plot_ips_calibration}
\title{Plot indifference points from the calibration task as a function of delay}
\usage{
plot_ips_calibration(tibb)
}
\description{
Individual indifference points, as well as group median and IQR. This plot
function is used in notebook 05_eda_grp.Rmd.
}
