% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_ll_choice_pctgs}
\alias{plot_ll_choice_pctgs}
\title{Plot larger-later choice percentage across frames}
\usage{
plot_ll_choice_pctgs(tibb, task)
}
\description{
This plot function is used in notebook 05_eda_grp.Rmd.
}
