% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_model_comparison_aggregate_bic}
\alias{plot_model_comparison_aggregate_bic}
\title{Plot results of model comparison based on aggregate BIC score}
\usage{
plot_model_comparison_aggregate_bic(tibb, fill_values)
}
\description{
Plot results of model comparison based on aggregate BIC score
}
