% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_model_comparison_bic_count}
\alias{plot_model_comparison_bic_count}
\title{Plot results of model comparison based on BIC counts}
\usage{
plot_model_comparison_bic_count(tibb, fill_values)
}
\description{
Plot results of model comparison based on BIC counts
}
