% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_obs_prd_performance}
\alias{plot_obs_prd_performance}
\title{Plot performance observed against performance predicted by the model}
\usage{
plot_obs_prd_performance(tibb, plotvar = "p_ll",
  key_frames_only = FALSE)
}
\description{
Plot performance observed against performance predicted by the model
}
