% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_par_values_pairs}
\alias{plot_par_values_pairs}
\title{Plot distribution of best-fitting parameter values}
\usage{
plot_par_values_pairs(tibb)
}
\description{
Plot distribution of best-fitting parameter values
}
