% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_rt_percentiles}
\alias{plot_rt_percentiles}
\title{Plot response time percentiles across frames}
\usage{
plot_rt_percentiles(tibb, task)
}
\description{
This plot function is used in notebook 05_eda_grp.Rmd.
}
