% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{relabel_mdl}
\alias{relabel_mdl}
\title{Relabel model levels}
\usage{
relabel_mdl(x)
}
\arguments{
\item{x}{character vector of to be relabeled model levels}
}
\description{
Relabel model levels
}
