% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{relabel_pmz}
\alias{relabel_pmz}
\title{Relabel parameterization levels}
\usage{
relabel_pmz(x)
}
\arguments{
\item{x}{character vector of to be relabeled parameterization levels}
}
\description{
Relabel parameterization levels
}
