% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{tidy_obs_prd_choice_rt}
\alias{tidy_obs_prd_choice_rt}
\title{Returns choices and response times observed and predicted for plotting in tidy, nested format}
\usage{
tidy_obs_prd_choice_rt(return_var = "obs", obs, model = "", pmz = "",
  parameters = NA)
}
\arguments{
\item{obs}{Tibble containing observed data}

\item{model}{Name of model (e.g. "DDM", "DFT_C", "DFT_RT")}

\item{pmz}{Parameterization (e.g. "date_delay_value_scaling)}

\item{parameters}{Vector (unnamed) with best-fitting parameter values}
}
\description{
Returns choices and response times observed and predicted for plotting in tidy, nested format
}
